sql recovery, timing module
